Remove string module
